+++ src/tools/MedEasy.RestObjects/IGenericPagedGetResponse.cs
@@ -5,11 +5,11 @@
         /// <summary>
         /// Total of items in the the result
         /// </summary>
-        int Total { get; }
+        long Total { get; }
 
         /// <summary>
         /// Number of items in the the result
         /// </summary>
-        int Count { get; }
+        long Count { get; }
     }
 }
